+++ bluez-async/src/lib.rs
@@ -7,21 +7,30 @@
 //!
 //! [`BluetoothSession']: struct.BluetoothSession.html
 
+mod adapter;
 mod bleuuid;
+mod characteristic;
+mod descriptor;
+mod device;
 mod events;
 mod introspect;
 mod messagestream;
+mod service;
 
+pub use self::adapter::AdapterId;
 pub use self::bleuuid::{uuid_from_u16, uuid_from_u32, BleUuid};
+pub use self::characteristic::{CharacteristicFlags, CharacteristicId, CharacteristicInfo};
+pub use self::descriptor::{DescriptorId, DescriptorInfo};
+pub use self::device::{DeviceId, DeviceInfo};
 pub use self::events::{AdapterEvent, BluetoothEvent, CharacteristicEvent, DeviceEvent};
 use self::introspect::IntrospectParse;
 use self::messagestream::MessageStream;
-use bitflags::bitflags;
+pub use self::service::{ServiceId, ServiceInfo};
 use bluez_generated::{
     OrgBluezAdapter1, OrgBluezDevice1, OrgBluezDevice1Properties, OrgBluezGattCharacteristic1,
     OrgBluezGattDescriptor1, OrgBluezGattService1, ORG_BLUEZ_ADAPTER1_NAME, ORG_BLUEZ_DEVICE1_NAME,
 };
-use dbus::arg::{cast, PropMap, RefArg, Variant};
+use dbus::arg::{PropMap, Variant};
 use dbus::nonblock::stdintf::org_freedesktop_dbus::{Introspectable, ObjectManager, Properties};
 use dbus::nonblock::{Proxy, SyncConnection};
 use dbus::Path;
@@ -29,7 +38,7 @@ use dbus_tokio::connection::IOResourceError;
 use futures::stream::{self, select_all, StreamExt};
 use futures::{FutureExt, Stream};
 use std::collections::HashMap;
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryInto;
 use std::fmt::{self, Debug, Display, Formatter};
 use std::future::Future;
 use std::str::FromStr;
@@ -76,204 +85,6 @@ pub enum SpawnError {
     Join(#[from] JoinError),
 }
 
-/// Opaque identifier for a Bluetooth adapter on the system.
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct AdapterId {
-    object_path: Path<'static>,
-}
-
-impl AdapterId {
-    fn new(object_path: &str) -> Self {
-        Self {
-            object_path: object_path.to_owned().into(),
-        }
-    }
-}
-
-impl Display for AdapterId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            self.object_path
-                .to_string()
-                .strip_prefix("/org/bluez/")
-                .ok_or(fmt::Error)?
-        )
-    }
-}
-
-/// Opaque identifier for a Bluetooth device which the system knows about. This includes a reference
-/// to which Bluetooth adapter it was discovered on, which means that any attempt to connect to it
-/// will also happen from that adapter (in case the system has more than one).
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct DeviceId {
-    object_path: Path<'static>,
-}
-
-impl DeviceId {
-    fn new(object_path: &str) -> Self {
-        Self {
-            object_path: object_path.to_owned().into(),
-        }
-    }
-
-    /// Get the ID of the Bluetooth adapter on which this device was discovered, e.g. `"hci0"`.
-    pub fn adapter(&self) -> AdapterId {
-        let index = self
-            .object_path
-            .rfind('/')
-            .expect("DeviceId object_path must contain a slash.");
-        AdapterId::new(&self.object_path[0..index])
-    }
-}
-
-impl From<DeviceId> for Path<'static> {
-    fn from(id: DeviceId) -> Self {
-        id.object_path
-    }
-}
-
-impl Display for DeviceId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            self.object_path
-                .to_string()
-                .strip_prefix("/org/bluez/")
-                .ok_or(fmt::Error)?
-        )
-    }
-}
-
-/// Opaque identifier for a GATT service on a Bluetooth device.
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct ServiceId {
-    object_path: Path<'static>,
-}
-
-impl ServiceId {
-    fn new(object_path: &str) -> Self {
-        Self {
-            object_path: object_path.to_owned().into(),
-        }
-    }
-
-    /// Get the ID of the device on which this service was advertised.
-    pub fn device(&self) -> DeviceId {
-        let index = self
-            .object_path
-            .rfind('/')
-            .expect("ServiceId object_path must contain a slash.");
-        DeviceId::new(&self.object_path[0..index])
-    }
-}
-
-impl From<ServiceId> for Path<'static> {
-    fn from(id: ServiceId) -> Self {
-        id.object_path
-    }
-}
-
-impl Display for ServiceId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            self.object_path
-                .to_string()
-                .strip_prefix("/org/bluez/")
-                .ok_or(fmt::Error)?
-        )
-    }
-}
-
-/// Opaque identifier for a GATT characteristic on a Bluetooth device.
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct CharacteristicId {
-    object_path: Path<'static>,
-}
-
-impl CharacteristicId {
-    fn new(object_path: &str) -> Self {
-        Self {
-            object_path: object_path.to_owned().into(),
-        }
-    }
-
-    /// Get the ID of the service on which this characteristic was advertised.
-    pub fn service(&self) -> ServiceId {
-        let index = self
-            .object_path
-            .rfind('/')
-            .expect("CharacteristicId object_path must contain a slash.");
-        ServiceId::new(&self.object_path[0..index])
-    }
-}
-
-impl From<CharacteristicId> for Path<'static> {
-    fn from(id: CharacteristicId) -> Self {
-        id.object_path
-    }
-}
-
-impl Display for CharacteristicId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            self.object_path
-                .to_string()
-                .strip_prefix("/org/bluez/")
-                .ok_or(fmt::Error)?
-        )
-    }
-}
-
-/// Opaque identifier for a GATT characteristic descriptor on a Bluetooth device.
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct DescriptorId {
-    object_path: Path<'static>,
-}
-
-impl DescriptorId {
-    #[cfg(test)]
-    fn new(object_path: &str) -> Self {
-        Self {
-            object_path: object_path.to_owned().into(),
-        }
-    }
-
-    /// Get the ID of the characteristic on which this descriptor was advertised.
-    pub fn characteristic(&self) -> CharacteristicId {
-        let index = self
-            .object_path
-            .rfind('/')
-            .expect("DescriptorId object_path must contain a slash.");
-        CharacteristicId::new(&self.object_path[0..index])
-    }
-}
-
-impl From<DescriptorId> for Path<'static> {
-    fn from(id: DescriptorId) -> Self {
-        id.object_path
-    }
-}
-
-impl Display for DescriptorId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            self.object_path
-                .to_string()
-                .strip_prefix("/org/bluez/")
-                .ok_or(fmt::Error)?
-        )
-    }
-}
-
 /// MAC address of a Bluetooth device.
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct MacAddress(String);
@@ -307,159 +118,6 @@ impl FromStr for MacAddress {
         }
         Ok(MacAddress(s.to_uppercase()))
     }
-}
-
-/// Information about a Bluetooth device which was discovered.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct DeviceInfo {
-    /// An opaque identifier for the device, including a reference to which adapter it was
-    /// discovered on. This can be used to connect to it.
-    pub id: DeviceId,
-    /// The MAC address of the device.
-    pub mac_address: MacAddress,
-    /// The human-readable name of the device, if available.
-    pub name: Option<String>,
-    /// The appearance of the device, as defined by GAP.
-    pub appearance: Option<u16>,
-    /// The GATT service UUIDs (if any) from the device's advertisement or service discovery.
-    ///
-    /// Note that service discovery only happens after a connection has been made to the device, but
-    /// BlueZ may cache the list of services after it is disconnected.
-    pub services: Vec<Uuid>,
-    /// Whether the device is currently paired with the adapter.
-    pub paired: bool,
-    /// Whether the device is currently connected to the adapter.
-    pub connected: bool,
-    /// The Received Signal Strength Indicator of the device advertisement or inquiry.
-    pub rssi: Option<i16>,
-    /// Manufacturer-specific advertisement data, if any. The keys are 'manufacturer IDs'.
-    pub manufacturer_data: HashMap<u16, Vec<u8>>,
-    /// The GATT service data from the device's advertisement, if any. This is a map from the
-    /// service UUID to its data.
-    pub service_data: HashMap<Uuid, Vec<u8>>,
-    /// Whether service discovery has finished for the device.
-    pub services_resolved: bool,
-}
-
-impl DeviceInfo {
-    fn from_properties(
-        id: DeviceId,
-        device_properties: OrgBluezDevice1Properties,
-    ) -> Result<DeviceInfo, BluetoothError> {
-        let mac_address = device_properties
-            .address()
-            .ok_or_else(|| BluetoothError::RequiredPropertyMissing("Address".to_string()))?;
-        let services = get_services(device_properties);
-        let manufacturer_data = get_manufacturer_data(device_properties).unwrap_or_default();
-        let service_data = get_service_data(device_properties).unwrap_or_default();
-
-        Ok(DeviceInfo {
-            id,
-            mac_address: MacAddress(mac_address.to_owned()),
-            name: device_properties.name().cloned(),
-            appearance: device_properties.appearance(),
-            services,
-            paired: device_properties
-                .paired()
-                .ok_or_else(|| BluetoothError::RequiredPropertyMissing("Paired".to_string()))?,
-            connected: device_properties
-                .connected()
-                .ok_or_else(|| BluetoothError::RequiredPropertyMissing("Connected".to_string()))?,
-            rssi: device_properties.rssi(),
-            manufacturer_data,
-            service_data,
-            services_resolved: device_properties.services_resolved().ok_or_else(|| {
-                BluetoothError::RequiredPropertyMissing("ServicesResolved".to_string())
-            })?,
-        })
-    }
-}
-
-/// Information about a GATT service on a Bluetooth device.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ServiceInfo {
-    /// An opaque identifier for the service on the device, including a reference to which adapter
-    /// it was discovered on.
-    pub id: ServiceId,
-    /// The 128-bit UUID of the service.
-    pub uuid: Uuid,
-    /// Whether this GATT service is a primary service.
-    pub primary: bool,
-}
-
-/// Information about a GATT characteristic on a Bluetooth device.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct CharacteristicInfo {
-    /// An opaque identifier for the characteristic on the device, including a reference to which
-    /// adapter it was discovered on.
-    pub id: CharacteristicId,
-    /// The 128-bit UUID of the characteristic.
-    pub uuid: Uuid,
-    /// The set of flags (a.k.a. properties) of the characteristic, defining how the characteristic
-    /// can be used.
-    pub flags: CharacteristicFlags,
-}
-
-bitflags! {
-    /// The set of flags (a.k.a. properties) of a characteristic, defining how the characteristic
-    /// can be used.
-    pub struct CharacteristicFlags: u16 {
-        const BROADCAST = 0x01;
-        const READ = 0x02;
-        const WRITE_WITHOUT_RESPONSE = 0x04;
-        const WRITE = 0x08;
-        const NOTIFY = 0x10;
-        const INDICATE = 0x20;
-        const SIGNED_WRITE = 0x40;
-        const EXTENDED_PROPERTIES = 0x80;
-        const RELIABLE_WRITE = 0x100;
-        const WRITABLE_AUXILIARIES = 0x200;
-        const ENCRYPT_READ = 0x400;
-        const ENCRYPT_WRITE = 0x800;
-        const ENCRYPT_AUTHENTICATED_READ = 0x1000;
-        const ENCRYPT_AUTHENTICATED_WRITE = 0x2000;
-        const AUTHORIZE = 0x4000;
-    }
-}
-
-impl TryFrom<Vec<String>> for CharacteristicFlags {
-    type Error = BluetoothError;
-
-    fn try_from(value: Vec<String>) -> Result<Self, BluetoothError> {
-        let mut flags = Self::empty();
-        for flag_string in value {
-            let flag = match flag_string.as_ref() {
-                "broadcast" => Self::BROADCAST,
-                "read" => Self::READ,
-                "write-without-response" => Self::WRITE_WITHOUT_RESPONSE,
-                "write" => Self::WRITE,
-                "notify" => Self::NOTIFY,
-                "indicate" => Self::INDICATE,
-                "authenticated-signed-write" => Self::SIGNED_WRITE,
-                "extended-properties" => Self::EXTENDED_PROPERTIES,
-                "reliable-write" => Self::RELIABLE_WRITE,
-                "writable-auxiliaries" => Self::WRITABLE_AUXILIARIES,
-                "encrypt-read" => Self::ENCRYPT_READ,
-                "encrypt-write" => Self::ENCRYPT_WRITE,
-                "encrypt-authenticated-read" => Self::ENCRYPT_AUTHENTICATED_READ,
-                "encrypt-authenticated-write" => Self::ENCRYPT_AUTHENTICATED_WRITE,
-                "authorize" => Self::AUTHORIZE,
-                _ => return Err(BluetoothError::FlagParseError(flag_string)),
-            };
-            flags.insert(flag);
-        }
-        Ok(flags)
-    }
-}
-
-/// Information about a GATT descriptor on a Bluetooth device.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct DescriptorInfo {
-    /// An opaque identifier for the descriptor on the device, including a reference to which
-    /// adapter it was discovered on.
-    pub id: DescriptorId,
-    /// The 128-bit UUID of the descriptor.
-    pub uuid: Uuid,
 }
 
 /// The type of transport to use for a scan.
@@ -1007,226 +665,5 @@ impl BluetoothSession {
         }
         Ok(select_all(message_streams)
             .flat_map(|message| stream::iter(BluetoothEvent::message_to_events(message))))
-    }
-}
-
-fn get_manufacturer_data(
-    device_properties: OrgBluezDevice1Properties,
-) -> Option<HashMap<u16, Vec<u8>>> {
-    Some(convert_manufacturer_data(
-        device_properties.manufacturer_data()?,
-    ))
-}
-
-pub(crate) fn convert_manufacturer_data(
-    data: &HashMap<u16, Variant<Box<dyn RefArg>>>,
-) -> HashMap<u16, Vec<u8>> {
-    data.iter()
-        .filter_map(|(&k, v)| {
-            if let Some(v) = cast::<Vec<u8>>(&v.0) {
-                Some((k, v.to_owned()))
-            } else {
-                log::warn!("Manufacturer data had wrong type: {:?}", &v.0);
-                None
-            }
-        })
-        .collect()
-}
-
-fn get_service_data(
-    device_properties: OrgBluezDevice1Properties,
-) -> Option<HashMap<Uuid, Vec<u8>>> {
-    // UUIDs don't get populated until we connect. Use:
-    //     "ServiceData": Variant(InternalDict { data: [
-    //         ("0000fe95-0000-1000-8000-00805f9b34fb", Variant([48, 88, 91, 5, 1, 23, 33, 215, 56, 193, 164, 40, 1, 0])
-    //     )], outer_sig: Signature("a{sv}") })
-    // instead.
-    Some(
-        device_properties
-            .service_data()?
-            .iter()
-            .filter_map(|(k, v)| match Uuid::parse_str(k) {
-                Ok(uuid) => {
-                    if let Some(v) = cast::<Vec<u8>>(&v.0) {
-                        Some((uuid, v.to_owned()))
-                    } else {
-                        log::warn!("Service data had wrong type: {:?}", &v.0);
-                        None
-                    }
-                }
-                Err(err) => {
-                    log::warn!("Error parsing service data UUID: {}", err);
-                    None
-                }
-            })
-            .collect(),
-    )
-}
-
-fn get_services(device_properties: OrgBluezDevice1Properties) -> Vec<Uuid> {
-    if let Some(uuids) = device_properties.uuids() {
-        uuids
-            .iter()
-            .filter_map(|uuid| {
-                Uuid::parse_str(uuid)
-                    .map_err(|err| {
-                        log::warn!("Error parsing service data UUID: {}", err);
-                        err
-                    })
-                    .ok()
-            })
-            .collect()
-    } else {
-        vec![]
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn device_adapter() {
-        let adapter_id = AdapterId::new("/org/bluez/hci0");
-        let device_id = DeviceId::new("/org/bluez/hci0/dev_11_22_33_44_55_66");
-        assert_eq!(device_id.adapter(), adapter_id);
-    }
-
-    #[test]
-    fn service_device() {
-        let device_id = DeviceId::new("/org/bluez/hci0/dev_11_22_33_44_55_66");
-        let service_id = ServiceId::new("/org/bluez/hci0/dev_11_22_33_44_55_66/service0022");
-        assert_eq!(service_id.device(), device_id);
-    }
-
-    #[test]
-    fn characteristic_service() {
-        let service_id = ServiceId::new("/org/bluez/hci0/dev_11_22_33_44_55_66/service0022");
-        let characteristic_id =
-            CharacteristicId::new("/org/bluez/hci0/dev_11_22_33_44_55_66/service0022/char0033");
-        assert_eq!(characteristic_id.service(), service_id);
-    }
-
-    #[test]
-    fn descriptor_characteristic() {
-        let characteristic_id =
-            CharacteristicId::new("/org/bluez/hci0/dev_11_22_33_44_55_66/service0022/char0033");
-        let descriptor_id = DescriptorId::new(
-            "/org/bluez/hci0/dev_11_22_33_44_55_66/service0022/char0033/desc0034",
-        );
-        assert_eq!(descriptor_id.characteristic(), characteristic_id);
-    }
-
-    #[test]
-    fn parse_flags() {
-        let flags: CharacteristicFlags = vec!["read".to_string(), "encrypt-write".to_string()]
-            .try_into()
-            .unwrap();
-        assert_eq!(
-            flags,
-            CharacteristicFlags::READ | CharacteristicFlags::ENCRYPT_WRITE
-        )
-    }
-
-    #[test]
-    fn parse_flags_fail() {
-        let flags: Result<CharacteristicFlags, BluetoothError> =
-            vec!["read".to_string(), "invalid flag".to_string()].try_into();
-        assert!(
-            matches!(flags, Err(BluetoothError::FlagParseError(string)) if string == "invalid flag")
-        );
-    }
-
-    #[test]
-    fn service_data() {
-        let uuid = uuid_from_u32(0x11223344);
-        let mut service_data: HashMap<String, Variant<Box<dyn RefArg>>> = HashMap::new();
-        service_data.insert(uuid.to_string(), Variant(Box::new(vec![1u8, 2, 3])));
-        let mut device_properties: HashMap<String, Variant<Box<dyn RefArg>>> = HashMap::new();
-        device_properties.insert("ServiceData".to_string(), Variant(Box::new(service_data)));
-
-        let mut expected_service_data = HashMap::new();
-        expected_service_data.insert(uuid, vec![1u8, 2, 3]);
-
-        assert_eq!(
-            get_service_data(OrgBluezDevice1Properties(&device_properties)),
-            Some(expected_service_data)
-        );
-    }
-
-    #[test]
-    fn manufacturer_data() {
-        let manufacturer_id = 0x1122;
-        let mut manufacturer_data: HashMap<u16, Variant<Box<dyn RefArg>>> = HashMap::new();
-        manufacturer_data.insert(manufacturer_id, Variant(Box::new(vec![1u8, 2, 3])));
-        let mut device_properties: HashMap<String, Variant<Box<dyn RefArg>>> = HashMap::new();
-        device_properties.insert(
-            "ManufacturerData".to_string(),
-            Variant(Box::new(manufacturer_data)),
-        );
-
-        let mut expected_manufacturer_data = HashMap::new();
-        expected_manufacturer_data.insert(manufacturer_id, vec![1u8, 2, 3]);
-
-        assert_eq!(
-            get_manufacturer_data(OrgBluezDevice1Properties(&device_properties)),
-            Some(expected_manufacturer_data)
-        );
-    }
-
-    #[test]
-    fn device_info_minimal() {
-        let id = DeviceId::new("/org/bluez/hci0/dev_11_22_33_44_55_66");
-        let mut device_properties: HashMap<String, Variant<Box<dyn RefArg>>> = HashMap::new();
-        device_properties.insert(
-            "Address".to_string(),
-            Variant(Box::new("00:11:22:33:44:55".to_string())),
-        );
-        device_properties.insert("Paired".to_string(), Variant(Box::new(false)));
-        device_properties.insert("Connected".to_string(), Variant(Box::new(false)));
-        device_properties.insert("ServicesResolved".to_string(), Variant(Box::new(false)));
-
-        let device =
-            DeviceInfo::from_properties(id.clone(), OrgBluezDevice1Properties(&device_properties))
-                .unwrap();
-        assert_eq!(
-            device,
-            DeviceInfo {
-                id,
-                mac_address: MacAddress("00:11:22:33:44:55".to_string()),
-                name: None,
-                appearance: None,
-                services: vec![],
-                paired: false,
-                connected: false,
-                rssi: None,
-                manufacturer_data: HashMap::new(),
-                service_data: HashMap::new(),
-                services_resolved: false,
-            }
-        )
-    }
-
-    #[test]
-    fn get_services_none() {
-        let device_properties: HashMap<String, Variant<Box<dyn RefArg>>> = HashMap::new();
-
-        assert_eq!(
-            get_services(OrgBluezDevice1Properties(&device_properties)),
-            vec![]
-        )
-    }
-
-    #[test]
-    fn get_services_some() {
-        let uuid = uuid_from_u32(0x11223344);
-        let uuids = vec![uuid.to_string()];
-        let mut device_properties: HashMap<String, Variant<Box<dyn RefArg>>> = HashMap::new();
-        device_properties.insert("UUIDs".to_string(), Variant(Box::new(uuids)));
-
-        assert_eq!(
-            get_services(OrgBluezDevice1Properties(&device_properties)),
-            vec![uuid]
-        )
     }
 }
